+++ tango/format.py
@@ -6,6 +6,7 @@ import json
 import logging
 import lzma
 import mmap
+import pathlib
 from abc import abstractmethod
 from os import PathLike
 from pathlib import Path
@@ -27,10 +28,12 @@ from typing import (
 import dill
 import xxhash
 
+from tango.common import DatasetDict, filename_is_safe
 from tango.common.aliases import PathOrStr
 from tango.common.exceptions import ConfigurationError
 from tango.common.logging import TangoLogger
 from tango.common.registrable import Registrable
+from tango.common.sqlite_sparse_sequence import SqliteSparseSequence
 
 T = TypeVar("T")
 
@@ -346,3 +349,38 @@ class JsonFormatIterator(Iterator[T], Generic[T]):
             self.f.close()
             self.f = None
             raise StopIteration()
+
+
+@Format.register("sqlite")
+class SqliteDictFormat(Format[DatasetDict]):
+    VERSION = 3
+
+    def write(self, artifact: DatasetDict, dir: Union[str, PathLike]):
+        dir = pathlib.Path(dir)
+        with gzip.open(dir / "metadata.dill.gz", "wb") as f:
+            dill.dump(artifact.metadata, f)
+        for split_name, split in artifact.splits.items():
+            filename = f"{split_name}.sqlite"
+            if not filename_is_safe(filename):
+                raise ValueError(f"{split_name} is not a valid name for a split.")
+            (dir / filename).unlink(missing_ok=True)
+            if isinstance(split, SqliteSparseSequence):
+                split.copy_to(dir / filename)
+            else:
+                sqlite = SqliteSparseSequence(dir / filename)
+                sqlite.extend(split)
+
+    def read(self, dir: Union[str, PathLike]) -> DatasetDict:
+        dir = pathlib.Path(dir)
+        with gzip.open(dir / "metadata.dill.gz", "rb") as f:
+            metadata = dill.load(f)
+        splits = {
+            filename.stem: SqliteSparseSequence(filename, read_only=True)
+            for filename in dir.glob("*.sqlite")
+        }
+        return DatasetDict(metadata=metadata, splits=splits)
+
+    def checksum(self, dir: PathOrStr) -> str:
+        # This is not trivial to implement because sqlite files can be different even if they contain the same
+        # data.
+        raise NotImplementedError()
